moved templates around fixed rendering
