--- interfaces/acados_template/acados_template/custom_update_template/__init__.py
@@ -1,1 +1,0 @@
-from .zoro_template.zoro_description import ZoroDescription
