+++ interfaces/acados_template/setup.py
@@ -85,8 +85,8 @@ setup(name='acados_template',
         'c_templates_tera/external_cost.in.h',
         'c_templates_tera/external_cost_e.in.h',
         'c_templates_tera/external_cost_0.in.h',
-        'custom_update_template/zoro_template/custom_update_function_zoro_template.in.c',
-        'custom_update_template/zoro_template/custom_update_function_zoro_template.in.h',
+        'custom_update_templates/custom_update_function_zoro_template.in.c',
+        'custom_update_templates/custom_update_function_zoro_template.in.h',
         'acados_layout.json',
         'acados_sim_layout.json',
         ]},
